Move files around till it feels right
