+++ src/app/(app)/[lang]/layout.tsx
@@ -1,4 +1,4 @@
-import { AppDataProvider } from './app-data-provider'
+import { AppDataProvider } from 'app/app-data-provider'
 
 export default function Layout({ params: { lang }, children }) {
   return <AppDataProvider lang={lang}>{children}</AppDataProvider>
